Added support for OSC, so that the application can be controlled from an Android phone, using ofxOsxReceiver. This is needed because on stage the computer with the software and the Kinects attached to it will be out of reach, and some things will need to be timed. (change color, start/stop rendering)
Note some changes will need to be made to the code of ofxOscReceiver to make it compatible with VS 2012, since otherwise you get errors when the 'unsafe' versions of sprintf, strcpy, etc are used. I'll fork this repo and commit my changes.

Also changed the 'togggle GUI' command to listen to the Tab key instead of Space, since the timeline was also listening to Space to start/stop playing, so this was interfering.